Pull Request Feedback.   Moved DiangosicSourceEventSource to its own file, other nits.
